Adding in submission guidelines.
This involved bringing some macros over from last year, which is OK. They might need cleanup, but...

The submissions pages have a more-complex-than-i-would-like include structure, but that was because different pages said different things in different places. For now, it's going to be what it is.

Some changes in how dates are handled... now, dates are in one place, as opposed to the replicated nightmare I had before. site.data.master.trackN is where each date goes for each track.