Fixing menu structure.
Updated data structure so we can have multiple levels of subheads. The absence of recursion in Liquid templates hurts.

Now using the SmartMenus extension for Bootstrap. BS3 got rid of multi-level menus, because they are considered bad practice on the WWW. We have a multi-level menu, and now have a CSS/JS library that makes it possible.

Included prospectus.