Added committees.
Somewhere, we should produce a guide for how this kind of data is managed by the SIGCSE team. Copy-pasting things out of PDFs is not ideal. I hope I have no introduced errors...

This introduces a data structure for the committees, and renders it. As usual, I have tried 3 different things to make anchors within a page work, and am unable to get the spacing right. Jumping to an anchor obscures the heading. However... it gets you there, which is step one.

Resolves #4