Updating registration info.
I pulled things forward from last year; not all of it is linked yet.

Registration rates are now on the site. Without early/late dates, I can't put those in... but, those are a few months away. And, I don't think registration is open... so, the information is now available.