Important dates, notifications
I have broken out both "Important Dates" and "Submission Guidelines" as separate links in the menu. They go to the same page. However, the user will look for those two things as separate ideas, and having them itemized in the menu makes sense to me.

I now know the dates we are going to alert authors, and have entered that info into the "database."

Closes #31